Renamed the inEach hook to inEachEnd and added an inEachBegin hook
