+++ packages/code-export-java/src/hooks/index.js
@@ -4,7 +4,8 @@ import beforeAll from './beforeAll'
 import beforeEach from './beforeEach'
 import afterEach from './afterEach'
 import afterAll from './afterAll'
-import inEach from './inEach'
+import inEachBegin from './inEachBegin'
+import inEachEnd from './inEachEnd'
 import methods from './declareMethods'
 
 const hooks = [
@@ -14,7 +15,8 @@ const hooks = [
   beforeEach,
   afterEach,
   afterAll,
-  inEach,
+  inEachBegin,
+  inEachEnd,
   methods,
 ]
 
@@ -31,6 +33,7 @@ export default {
   beforeEach,
   afterEach,
   afterAll,
-  inEach,
+  inEachBegin,
+  inEachEnd,
   methods,
 }
